refactor: workspace split
