+++ .goreleaser.yaml
@@ -4,7 +4,7 @@
 version: 2
 
 # Project name should ideally match the project name in your Cargo.toml.
-project_name: "pfs_rs"
+project_name: "pfs-rs"
 
 before:
   hooks:
@@ -18,6 +18,9 @@ builds:
     builder: rust
     targets:
       - x86_64-unknown-linux-gnu
+    flags:
+      - --release
+      - -p=pfs-rs
   # - id: linux-amd64-musl
   #   builder: rust
   #   targets:
@@ -26,6 +29,9 @@ builds:
     builder: rust
     targets:
       - aarch64-unknown-linux-gnu
+    flags:
+      - --release
+      - -p=pfs-rs
   # - id: linux-arm64-musl
   #   builder: rust
   #   targets:
@@ -36,6 +42,9 @@ builds:
     # command: build
     targets:
       - x86_64-pc-windows-gnu
+    flags:
+      - --release
+      - -p=pfs-rs
 
 # Creates archives for each target.
 archives:
