+++ Cargo.toml
@@ -1,15 +1,14 @@
-[package]
-name = "pfs_rs"
-version = "0.1.4"
-edition = "2024"
+[workspace]
+members = ["pfs-rs", "pf8"]
+resolver = "2"
 
 [profile.release]
 lto = true
 strip = true
 
-[dependencies]
+[workspace.dependencies]
 anyhow = "1.0.98"
-clap = { version = "4.5.42", features = ["derive"] }
+clap = { version = "4.5.43", features = ["derive"] }
 env_logger = "0.11.8"
 glob = "0.3.2"
 log = "0.4.27"
@@ -18,6 +17,5 @@ sha1 = "0.10.6"
 walkdir = "2.5.0"
 tabled = "0.20.0"
 human_bytes = { version = "0.4.3", default-features = false }
+tempfile = { version = "3.20.0" }
 
-[dev-dependencies]
-tempfile = "3.20.0"
